requirements.txt
flask
flask-session
langgraph
google-generativeai
faiss-cpu
langchain
langchain-text-splitters
pymupdf
python-dotenv
pytest
